move: DashboardType 패키지 이동(#23)
