+++ src/main/java/com/dnd/namuiwiki/domain/question/entity/Question.java
@@ -2,7 +2,7 @@ package com.dnd.namuiwiki.domain.question.entity;
 
 import com.dnd.namuiwiki.domain.option.entity.Option;
 import com.dnd.namuiwiki.domain.question.type.QuestionType;
-import com.dnd.namuiwiki.domain.statistic.type.DashboardType;
+import com.dnd.namuiwiki.domain.dashboard.type.DashboardType;
 import lombok.Builder;
 import lombok.Getter;
 import org.springframework.data.annotation.Id;
