+++ src/main/java/com/dnd/namuiwiki/domain/statistic/model/AverageStatistic.java
@@ -4,7 +4,7 @@ import com.dnd.namuiwiki.common.exception.ApplicationErrorException;
 import com.dnd.namuiwiki.common.exception.ApplicationErrorType;
 import com.dnd.namuiwiki.domain.option.entity.Option;
 import com.dnd.namuiwiki.domain.question.entity.Question;
-import com.dnd.namuiwiki.domain.statistic.type.DashboardType;
+import com.dnd.namuiwiki.domain.dashboard.type.DashboardType;
 import com.dnd.namuiwiki.domain.survey.model.entity.Survey;
 
 public class AverageStatistic extends Statistic {
