+++ src/main/java/com/dnd/namuiwiki/domain/statistic/model/Statistic.java
@@ -3,7 +3,7 @@ package com.dnd.namuiwiki.domain.statistic.model;
 import com.dnd.namuiwiki.common.exception.ApplicationErrorException;
 import com.dnd.namuiwiki.common.exception.ApplicationErrorType;
 import com.dnd.namuiwiki.domain.question.entity.Question;
-import com.dnd.namuiwiki.domain.statistic.type.DashboardType;
+import com.dnd.namuiwiki.domain.dashboard.type.DashboardType;
 import com.dnd.namuiwiki.domain.statistic.type.StatisticsType;
 import com.dnd.namuiwiki.domain.survey.model.entity.Survey;
 import lombok.AllArgsConstructor;
